+++ CHANGELOG.md
@@ -15,9 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ItemTransform` class, which describes item-wise dataset `Transform`s (<https://github.com/openvinotoolkit/datumaro/pull/297>)
 - `keep-empty` export parameter in VOC format (<https://github.com/openvinotoolkit/datumaro/pull/297>)
 - A base class for dataset validation plugins (<https://github.com/openvinotoolkit/datumaro/pull/299>)
+- Partial support for the Open Images format;
+  only reading is supported, and only images and image-level labels can be read
+  (<https://github.com/openvinotoolkit/datumaro/pull/291>).
 
 ### Changed
 - Tensorflow AVX check is made optional in API and is disabled by default (<https://github.com/openvinotoolkit/datumaro/pull/305>)
+- Extensions for images in ImageNet_txt are now mandatory (<https://github.com/openvinotoolkit/datumaro/pull/302>)
+- Several dependencies now have lower bounds (<https://github.com/openvinotoolkit/datumaro/pull/308>)
 
 ### Deprecated
 -